+++ R/tar_branch_index.R
@@ -27,6 +27,7 @@
 #' }
 tar_branch_index <- function(names, store = targets::tar_config_get("store")) {
   tar_assert_chr(names)
+  tar_assert_store(store = store)
   tar_assert_path(path_meta(path_store = store))
   meta <- meta_init(path_store = store)
   meta <- tibble::as_tibble(meta$database$read_condensed_data())