+++ R/tar_process.R
@@ -39,7 +39,9 @@ tar_process <- function(
   names = NULL,
   store = targets::tar_config_get("store")
 ) {
-  tar_assert_path(path_process(path_store = store))
+  tar_assert_scalar(store)
+  tar_assert_chr(store)
+  tar_assert_nzchar(store)
   out <- tibble::as_tibble(process_init(path_store = store)$read_process())
   names_quosure <- rlang::enquo(names)
   names <- tar_tidyselect_eval(names_quosure, out$name)