+++ R/tar_read.R
@@ -31,6 +31,7 @@ tar_read <- function(
   meta = tar_meta(store = store),
   store = targets::tar_config_get("store")
 ) {
+  tar_assert_store(store = store)
   force(meta)
   name <- tar_deparse_language(substitute(name))
   tar_read_raw(